clean up warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7/bzflag@654 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -5850,7 +5850,7 @@ static void parse(int argc, char **argv)
         cerr << "argument expected for " << argv[i] << endl;
         usage(argv[0]);
       }
-      idlekickthresh = atoi(argv[i]);
+      idlekickthresh = (float) atoi(argv[i]);
     }
     else {
       cerr << "bad argument " << argv[i] << endl;
